Dynamic model/requirements.txt
mediapipe
tensorflow
opencv-python
numpy
google-generativeai
streamlit
googletrans
playsound
gtts
scipy